requirements.txt
amqp==2.6.1
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
Django==3.2.5
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
djangorestframework==3.12.4
docker==5.0.0
flower==0.9.7
gprof2dot==2021.2.21
humanize==3.5.0
idna==2.10
Jinja2==3.0.1
kombu==4.6.11
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.58.0
